Moved Socket io to new component and reversed mongoose find controller
