--- client/src/components/Post.js
@@ -1,12 +1,0 @@
-import React, { Component } from 'react';
-import API from '../utils/API'
-
-class Post extends Component{
-    render(){
-        return(
-            <h1>Post here</h1>
-        )
-    }
-}
-
-export default Post;
