+++ client/src/components/Posts/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Posts";
