+++ client/src/components/Profile/Profile.js
@@ -4,8 +4,7 @@ import withAuth from './../withAuth';
 import API from '../../utils/API';
 import { Link } from 'react-router-dom';
 import "./Profile.css";
-import { Panel, Collapse } from "react-bootstrap";
-import socketIOClient from 'socket.io-client'
+import Posts from "./../Posts";
 const Auth = new AuthService();
 
 class Profile extends Component {
@@ -13,14 +12,8 @@ class Profile extends Component {
     constructor() {
         super()
         this.state = {
-            endpoint: "http://localhost:3001",
-            email: "",
             sideBar: true,
             open: false,
-            results: [],
-            title: "",
-            category: "",
-            content: ""
         }
     }
 
@@ -30,8 +23,6 @@ class Profile extends Component {
                 email: res.data.email
             })
         });
-
-        this.getPosts();
     }
 
     toggleSide = () => {
@@ -45,62 +36,7 @@ class Profile extends Component {
         this.props.history.replace('/login');
     }
 
-    handleInputChange = event => {
-        const { name, value } = event.target;
-        this.setState({
-            [name]: value
-        });
-    };
-
-    handleFormSubmit = event => {
-        event.preventDefault();
-        if (this.state.title && this.state.content) {
-            API.create({
-                title: this.state.title,
-                category: this.state.category,
-                content: this.state.content
-            })
-                .then(res => {
-                    this.getPosts()
-                    console.log(res);
-                }
-
-                ).then(this.setState({
-                    title: "",
-                    category: "",
-                    content: ""
-                }))
-                .catch(err => console.log(err));
-        }
-    };
-
-    getPosts = () => {
-        API.findPost()
-            .then(res => this.setState({
-                results: res.data
-            }))
-
-            .then(() => { this.createPost() })
-            .catch(err => console.log(err));
-    }
-
-    createPost = () => {
-        const socket = socketIOClient(this.state.endpoint);
-        socket.emit('post', this.state.results);
-      }
-
     render() {
-
-        const socket = socketIOClient(this.state.endpoint)
-
-        socket.on('post', results => {
-            this.setState({
-              results : results
-            });
-            console.log("message from socket", this.state.results);
-          });
-
-
         return (
             <div className="Profile">
                 <div className="row">
@@ -112,58 +48,7 @@ class Profile extends Component {
                         <Link to="/">Go home</Link>
                     </div>
                     <div className={this.state.sideBar ? "col-md-8 text-center" : "col-md-8 text-center mainFull"}>
-                        <h2>Whistles</h2>
-                        <Panel>
-                            <Panel.Heading>
-                                <h4 onClick={() => this.setState({ open: !this.state.open })}>Blow the Whistle</h4>
-                                <Collapse in={this.state.open}>
-                                    <form className="form-horizontal">
-                                        <div className="form-group">
-                                            <label htmlFor="Title" className="col-sm-2 control-label">Title</label>
-                                            <div className="col-sm-10">
-                                                <input value={this.state.title}
-                                                    onChange={this.handleInputChange}
-                                                    name="title" type="text" className="form-control" id="Title" placeholder="Title" />
-                                            </div>
-                                        </div>
-                                        <div className="form-group">
-                                            <label htmlFor="category" className="col-sm-2 control-label">Category</label>
-                                            <div className="col-sm-10">
-                                                <input value={this.state.category}
-                                                    onChange={this.handleInputChange}
-                                                    name="category" type="text" className="form-control" id="category" placeholder="Category" />
-                                            </div>
-                                        </div>
-                                        <div className="form-group">
-                                            <label htmlFor="content" className="col-sm-2 control-label">Content</label>
-                                            <div className="col-sm-10">
-                                                <textarea value={this.state.content}
-                                                    onChange={this.handleInputChange}
-                                                    name="content" id="content" className="form-control" rows="3" placeholder="Content"></textarea>
-                                            </div>
-                                        </div>
-                                        <div className="form-group">
-                                            <div className="col-sm-12">
-                                                <button onClick={this.handleFormSubmit} type="submit" className="btn btn-default">Submit</button>
-                                            </div>
-                                        </div>
-                                    </form>
-                                </Collapse>
-                            </Panel.Heading>
-                            <Panel.Body>
-                                {this.state.results.length === 0 ? "Nothing Posted" :
-                                    this.state.results.map((key, index) =>
-
-                                        <Panel key={index}>
-                                            <Panel.Body>
-                                                <p>Title: {key.title} Category: {key.category}</p>
-                                                <div id="panelContent">Content: {key.content}</div>
-                                            </Panel.Body>
-                                        </Panel>
-
-                                    )}
-                            </Panel.Body>
-                        </Panel>
+                        <Posts/>
                     </div>
 
                 </div>
