+++ client/src/index.js
@@ -12,7 +12,6 @@ import Home from './components/Home';
 import Login from './components/Login';
 import Profile from './components/Profile';
 import Signup from './components/Signup';
-import Post from './components/Post';
 import About from './components/About';
 import Live from './components/Live';
 
@@ -27,7 +26,6 @@ ReactDOM.render(
             <Route exact path="/login" component={Login} />
             <Route exact path="/signup" component={Signup} />
             <Route exact path="/profile/:id" component={Profile} />
-            <Route excat path="/post"component={Post}/>
             <Route exact path="/about" component={About} />
             <Route exact path="/live" component={Live} />
